+++ package/tsdown.config.ts
@@ -3,16 +3,13 @@ import { cjsPreset, dtsPreset, esmPreset } from "@apst/tsdown/presets";
 
 export default defineConfig(
     {
-        entry: {
+        entry: [
             // public
-            rolldown: "./src/rolldown.ts",
-            vite: "./src/vite.ts",
+            "./src/*.ts",
             // internal
-            "plugins/config": "./src/plugins/config.ts",
-            "plugins/transform": "./src/plugins/transform.ts",
-            "plugins/minify": "./src/plugins/minify.ts",
-            "functions/tsconfig": "./src/functions/tsconfig.ts",
-        },
+            "./src/plugins/*.ts",
+            "./src/functions/tsconfig.ts",
+        ],
         unbundle: true,
     },
     [